Fold dexterous lump review fixes into the reorient manager layer
P5-owned share of the lump review-response commits:

- Relocate the shared reset/rotation helpers into reorient/mdp/events.py
  and EpisodeErrorRecorder into reorient/mdp/rewards.py, removing the
  core/utils.py module (S17, from 428297c2).
- Default the reorient manager cfgs to newton_mjwarp and source the
  hand name lists from the robot assets (from 3f9ce324, 34676102).
- Repoint the Direct env to the relocated helpers.
