+++ source/isaaclab_tasks/isaaclab_tasks/core/reorient/config/shadow_hand/shadow_hand_manager_env_cfg.py
@@ -23,13 +23,13 @@ from isaaclab_tasks.core.reorient.config.shadow_hand.shadow_hand_common import (
     GOAL_OBJECT_CFG,
     OBJECT_CFG,
     ROBOT_CFG,
-    SHADOW_ACTUATED_JOINT_NAMES,
-    SHADOW_FINGERTIP_BODY_NAMES,
     ObjectCfg,
     PhysicsCfg,
 )
 from isaaclab_tasks.core.reorient.reorient_common import GOAL_MARKER_POSITION, IN_HAND_POS_OFFSET
 from isaaclab_tasks.utils import PresetCfg
+
+from isaaclab_assets.robots.shadow_hand import SHADOW_ACTUATED_JOINT_NAMES, SHADOW_FINGERTIP_BODY_NAMES
 
 # ---------------------------------- state task ----------------------------------
 
@@ -59,7 +59,7 @@ class ShadowHandManagerSceneCfg(PresetCfg):
     newton_mjwarp = _ShadowHandManagerSceneCfg(clone_in_fabric=False)
     ovphysx = physx
     newton_kamino = newton_mjwarp
-    default = physx
+    default = newton_mjwarp
 
 
 @configclass
