+++ source/isaaclab_tasks/isaaclab_tasks/core/reorient/reorient_direct_env.py
@@ -20,9 +20,9 @@ from isaaclab.sensors import JointWrenchSensor, JointWrenchSensorCfg
 from isaaclab.sim.spawners.from_files import GroundPlaneCfg, spawn_ground_plane
 from isaaclab.utils.math import quat_conjugate, quat_mul, sample_uniform, saturate, scale_transform, unscale_transform
 
-from isaaclab_tasks.core.reorient.mdp.rewards import evaluate_reorient_success, reorient_reward
+from isaaclab_tasks.core.reorient.mdp.events import randomize_rotation, sample_joint_positions_within_limits
+from isaaclab_tasks.core.reorient.mdp.rewards import EpisodeErrorRecorder, evaluate_reorient_success, reorient_reward
 from isaaclab_tasks.core.reorient.reorient_common import GOAL_MARKER_POSITION, IN_HAND_POS_OFFSET
-from isaaclab_tasks.core.utils import EpisodeErrorRecorder, randomize_rotation, sample_joint_positions_within_limits
 
 if TYPE_CHECKING:
     from isaaclab_tasks.core.reorient.config.allegro_hand.allegro_hand_direct_env_cfg import AllegroHandEnvCfg
